requirements (1) (1).txt
discord.py==2.4.0
aiohttp==3.9.1
qrcode==7.4.2
Pillow==10.1.0
chat-exporter==2.6.0
sqlite3